wsd: config: move RuntimeConstant to ConfigUtil
It's only used by the config logic.
